refactor(registry): mechanical split registry_client tests (wave11 step2) (#629)
* refactor(registry): mechanical split registry_client tests into modules (wave11 step2)

* docs(split): add wave11 step2 review pack + move map + gate
